Create SECURITY.md
# Security Policy

 ## Supported Versions
We provide security updates for the following versions:

| Version | Supported                  |
| -------       | -------------------------------- |
| 1.x.x      | :white_check_mark:  |
| < 1.x.x   | :x:                                 |

## Reporting a Vulnerability
If you discover a security vulnerability in Gabillon-100, we strongly encourage responsible disclosure. Please do **not** report vulnerabilities in public issues or pull requests. Instead, follow the steps below:

### 1. Contact Us
Please send an email to **gaabrielogar100@gmail.com* with the following information:
- A description of the vulnerability.
- Steps to reproduce the issue.
- The affected version.
- Any potential fixes or workarounds.

### 2. Response Time
We will acknowledge your report within **48 hours** and work to resolve the issue as quickly as possible. You will be kept updated on our progress.

### 3. Bounty Program
If the vulnerability is significant and affects production systems, you may be eligible for a security bounty. Please provide details and references for eligibility.

## Security Best Practices
We recommend the following best practices when using Gabillon-100:
- Keep dependencies up to date.
- Use secure connections (TLS/SSL).
- Regularly update your instance to the latest version.

Thank you for helping us keep  Gabillon-100 secure!